Added lazy conditional evaluation.
Added another pass in splTranslator that converts A || B in if and while
statements into nested ifs so that B is only evaluated if A is false.

Also had to change the way loops are de-sugared in simplifier.ml.
It used to be:

// Loop location:
foo_loop()

proc foo_loop(..):
  prebody
  if (loop condition):
    postbody
    foo_loop()

Now, I've changed it to:
// Loop location:
prebody
foo_loop()

proc foo_loop(..):
  if (loop condition):
    postbody
    prebody
    foo_loop()

The reason for this change is that to remove the A || B from the loop condition,
it was necessary to introduce a temporary boolean variable tmp_cond that we set
to A || B using nested ifs in the prebody. But we also needed to add an invariant
saying tmp_cond <==> A || B. But prebody needs to be executed before the call to
foo_loop(), since invariants are translated to preconditions.